+++ packages/content/content.ts
@@ -1,5 +1,4 @@
 export * from './src/api-endpoints';
 export * from './src/content-store-service';
-export * from './src/engine-client';
 export * from './src/navigation-service';
 export * from './src/url-transformation-service';