+++ packages/engine/src/api-endpoints.ts
@@ -1,7 +1,0 @@
-export const GET_ITEM_URL = '/api/1/site/content_store/item.json';
-export const GET_DESCRIPTOR = '/api/1/site/content_store/descriptor.json';
-export const GET_CHILDREN = '/api/1/site/content_store/children.json';
-export const GET_TREE = '/api/1/site/content_store/tree.json';
-export const GET_NAV_TREE = '/api/1/site/navigation/tree.json';
-export const GET_BREADCRUMB = '/api/1/site/navigation/breadcrumb.json';
-export const TRANSFORM_URL = '/api/1/site/url/transform.json';