+++ packages/search/search.ts
@@ -1,3 +1,3 @@
 export * from './src/query';
 export * from './src/solr-query';
-export * from './src/solr-service';
+export * from './src/search-service';